Reviewed-on: http://172.16.0.19:8000/P_UJA/BOLSA_EMPLEO/pulls/326

+++ src/es/ujaen/uvirtual/modulo/bolsaempleo/modelo/ModeloMerito.java
@@ -185,7 +185,7 @@ public class ModeloMerito {
 			int parameterIndex = 1;
 			stmt.setInt(parameterIndex++, merito.getItemBaremacion().getCodNum());
 			stmt.setInt(parameterIndex++, usuarioUpdate.getCodNum());
-			stmt.setFloat(parameterIndex++, merito.getValor());
+			stmt.setDouble(parameterIndex++, merito.getValor());
 			stmt.setString(parameterIndex++, merito.getDescripcion());
 			stmt.setString(parameterIndex++, merito.getObservacion());
 			stmt.setBinaryStream(parameterIndex++, merito.getArchivo());
@@ -224,7 +224,7 @@ public class ModeloMerito {
 		try (Connection conexion = ConexionUvirtual.obtenerInstancia(); PreparedStatement stmt = conexion.prepareStatement(consulta)) {
 			int parameterIndex = 1;
 			stmt.setInt(parameterIndex++, merito.getItemBaremacion().getCodNum());
-			stmt.setFloat(parameterIndex++, merito.getValor());
+			stmt.setDouble(parameterIndex++, merito.getValor());
 			stmt.setInt(parameterIndex++, merito.getCodNum());
 			stmt.setString(parameterIndex++, usuarioUpdate.getCodCuenta());
 			stmt.executeUpdate();
@@ -360,7 +360,7 @@ public class ModeloMerito {
 		Merito mer = new Merito();		
 		mer.setCodNum(rs.getInt("CODNUM"));
 		mer.setItemBaremacion(modeloBar.getItemBaremacionById(rs.getInt("BEPITE_CODNUM")));		
-		mer.setValor(rs.getFloat("VALOR"));
+		mer.setValor(rs.getDouble("VALOR"));
 		mer.setDescripcion(rs.getString("DESCRIPCION"));
 		mer.setObservacion(rs.getString("OBSERVACION"));
